Add global setting 'DeleteAfterSuccess' for all tests

+++ ATL.test/CatalogDataReaders/CUETest.cs
@@ -27,7 +27,7 @@ namespace ATL.test
                 Assert.IsNotInstanceOfType(theReader, typeof(CatalogDataReaders.BinaryLogic.DummyReader));
                 Assert.AreEqual("Nintendo Sound Team", theReader.Artist);
                 Assert.AreEqual("Duck Hunt", theReader.Title);
-                Assert.AreEqual("GENRE Game" + Settings.InternalValueSeparator + "DATE 1984", theReader.Comments);
+                Assert.AreEqual("GENRE Game" + ATL.Settings.InternalValueSeparator + "DATE 1984", theReader.Comments);
 
                 Assert.AreEqual(2, theReader.Tracks.Count);
                 Assert.AreEqual("Title Screen", theReader.Tracks[0].Title);
@@ -39,7 +39,7 @@ namespace ATL.test
             }
             finally
             {
-                File.Delete(testFileLocation);
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
             }
         }
 
@@ -55,22 +55,22 @@ namespace ATL.test
                 Assert.IsNotInstanceOfType(theReader, typeof(CatalogDataReaders.BinaryLogic.DummyReader));
                 Assert.AreEqual("Faithless", theReader.Artist);
                 Assert.AreEqual("Fake sample Cuesheet", theReader.Title);
-                Assert.AreEqual("GENRE Electronica" + Settings.InternalValueSeparator + "DATE 1998", theReader.Comments);
+                Assert.AreEqual("GENRE Electronica" + ATL.Settings.InternalValueSeparator + "DATE 1998", theReader.Comments);
 
                 Assert.AreEqual(8, theReader.Tracks.Count);
                 Assert.AreEqual("Reverence", theReader.Tracks[0].Title);
                 Assert.AreEqual("Faithless1", theReader.Tracks[0].Artist);
                 Assert.AreEqual(403, theReader.Tracks[0].Duration);
-                Assert.AreEqual("comment11" + Settings.InternalValueSeparator + "comment12", theReader.Tracks[0].Comment);
+                Assert.AreEqual("comment11" + ATL.Settings.InternalValueSeparator + "comment12", theReader.Tracks[0].Comment);
 
                 Assert.AreEqual("She's My Baby", theReader.Tracks[1].Title);
                 Assert.AreEqual("Faithless", theReader.Tracks[1].Artist);
                 Assert.AreEqual(253, theReader.Tracks[1].Duration);
-                Assert.AreEqual("comment21" + Settings.InternalValueSeparator + "comment22", theReader.Tracks[1].Comment);
+                Assert.AreEqual("comment21" + ATL.Settings.InternalValueSeparator + "comment22", theReader.Tracks[1].Comment);
             }
             finally
             {
-                File.Delete(testFileLocation);
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
             }
         }
