+++ ATL.test/Playlist/B4SIO.cs
@@ -26,7 +26,7 @@ namespace ATL.test.IO.Playlist
             }
             finally
             {
-                File.Delete(testFileLocation);
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
             }
         }
 
@@ -123,7 +123,7 @@ namespace ATL.test.IO.Playlist
             }
             finally
             {
-                File.Delete(testFileLocation);
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
             }
         }
